Disable local explorer node temporarily
The latest docker image from hiro apparently has some issues
related to switching to custom network (fails with HTTP 500),
so I'm disabling launching it as part of the environment for
now. Instead an explicit explorer instance will be launched on
the staging environments. This can be reverted once the issue is
resolved upstream.